homework2/pyproject.toml
[tool.poetry]
name = "homework2"
version = "0.1.0"
description = ""
authors = ["Firas Al Wadhahi <Firas.Wadhahi@rihal.om>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
grpcio-tools = "^1.60.0"
grpcio = "^1.60.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
